Modularize LiveMode
