+++ src/le_beta_vis/frontend/livemode/widgets/DetectedClusterCollectionWidget.py
@@ -1,9 +1,10 @@
-"""Animated snake-path grid of cluster thumbnails.
+"""Right panel widget for the Live Mode screensaver.
 
-Each cell is a ``QLabel`` positioned absolutely within the grid widget.
+Displays all recently detected clusters as an animated snake-path grid.
+Each cell is a ``QLabel`` positioned absolutely within the widget.
 On each advance step, cells animate to their new positions using
 ``QPropertyAnimation`` grouped in a ``QParallelAnimationGroup``.
-Cell size is computed dynamically from the widget's available space.
+Cell size is computed dynamically from the available space.
 """
 
 from typing import List, Optional
@@ -11,22 +12,21 @@ from typing import List, Optional
 import numpy as np
 from PySide6.QtCore import (
     QEasingCurve,
+    QParallelAnimationGroup,
     QPoint,
+    QPropertyAnimation,
     QSize,
     Qt,
     Signal,
 )
 from PySide6.QtGui import QPixmap, QResizeEvent
 from PySide6.QtWidgets import QLabel, QWidget
-from PySide6.QtCore import QPropertyAnimation, QParallelAnimationGroup
 
 from le_beta_vis.common.Cluster import Cluster
 from le_beta_vis.frontend.fitsconverters.interface import Colormap
-from le_beta_vis.frontend.widgets.EnergyClusterWidget import (
-    EnergyClusterWidget,
-)
+from le_beta_vis.frontend.widgets.EnergyClusterWidget import EnergyClusterWidget
 
-from .LiveModeViewModel import LiveModeViewModel
+from ..LiveModeViewModel import LiveModeViewModel
 
 
 class _ThumbnailCell(QLabel):
@@ -59,37 +59,36 @@ class _ThumbnailCell(QLabel):
                 pm.fill(Qt.black)
                 self.setPixmap(pm)
             return
-        pixmap = EnergyClusterWidget.to_pixmap(
-            cluster.data,
-            colormap,
-            cell_size,
-        )
-        self.setPixmap(pixmap)
+        self.setPixmap(EnergyClusterWidget.to_pixmap(cluster.data, colormap, cell_size))
 
 
-class _ThumbnailGridWidget(QWidget):
-    """Right-side thumbnail grid for the Live Mode screensaver.
+class DetectedClusterCollectionWidget(QWidget):
+    """Animated snake-path grid panel showing all recently detected clusters.
 
-    Displays ``rows * cols`` thumbnail cells arranged in a snake
-    path.  Animation is powered by ``QPropertyAnimation``.
+    Displays ``rows * cols`` thumbnail cells arranged in a snake path.
+    Animation is powered by ``QPropertyAnimation``.
 
     Signals:
         advanceFinished: Emitted when the advance animation completes.
+
+    Args:
+        vm: The Live Mode ViewModel.
+        parent: Optional parent widget.
     """
 
     advanceFinished = Signal()
 
     def __init__(
         self,
-        viewModel: LiveModeViewModel,
+        vm: LiveModeViewModel,
         parent: Optional[QWidget] = None,
     ) -> None:
         super().__init__(parent)
-        self._vm = viewModel
-        rows, cols = viewModel.grid_shape
+        self._vm = vm
+        rows, cols = vm.grid_shape
         self._rows = rows
         self._cols = cols
-        self._spacing = viewModel.grid_spacing
+        self._spacing = vm.grid_spacing
         self._cell_size = 1
         self._cells: List[_ThumbnailCell] = []
         self._anim_group: Optional[QParallelAnimationGroup] = None
@@ -98,7 +97,7 @@ class _ThumbnailGridWidget(QWidget):
     # --- Public ---
 
     def populate(self, grid: List[Optional[Cluster]]) -> None:
-        """Renders grid state into cell pixmaps without animation.
+        """Render grid state into cell pixmaps without animation.
 
         Args:
             grid: List of clusters matching the grid capacity.
@@ -108,15 +107,6 @@ class _ThumbnailGridWidget(QWidget):
         for i, cell in enumerate(self._cells):
             cluster = grid[i] if i < len(grid) else None
             cell.setClusterPixmap(cluster, colormap, self._cell_size, empty_pm)
-
-    @staticmethod
-    def _makeEmptyPixmap(
-        colormap: Optional[Colormap],
-        cell_size: int,
-    ) -> QPixmap:
-        """Generate a pixmap representing the colormap's zero value."""
-        zero_data = np.zeros((1, 1), dtype=np.float32)
-        return EnergyClusterWidget.to_pixmap(zero_data, colormap, cell_size)
 
     def animateAdvance(
         self,
@@ -165,28 +155,29 @@ class _ThumbnailGridWidget(QWidget):
 
     def _createCells(self) -> None:
         """Instantiate all thumbnail cell widgets."""
-        total = self._rows * self._cols
-        for _ in range(total):
+        for _ in range(self._rows * self._cols):
             cell = _ThumbnailCell(self)
             cell.setFixedSize(QSize(1, 1))
             self._cells.append(cell)
 
+    @staticmethod
+    def _makeEmptyPixmap(colormap: Optional[Colormap], cell_size: int) -> QPixmap:
+        """Generate a pixmap representing the colormap's zero value."""
+        return EnergyClusterWidget.to_pixmap(
+            np.zeros((1, 1), dtype=np.float32), colormap, cell_size
+        )
+
     def _computeCellSize(self) -> int:
         """Derives cell side length from available widget space."""
-        w = self.width()
-        h = self.height()
         if self._cols <= 0 or self._rows <= 0:
             return 1
-        cell_w = (w - (self._cols - 1) * self._spacing) // self._cols
-        cell_h = (h - (self._rows - 1) * self._spacing) // self._rows
+        cell_w = (self.width() - (self._cols - 1) * self._spacing) // self._cols
+        cell_h = (self.height() - (self._rows - 1) * self._spacing) // self._rows
         return max(1, min(cell_w, cell_h))
 
     def _computeVerticalOffset(self) -> int:
         """Returns top margin needed to vertically center the grid."""
-        used_h = (
-            self._rows * self._cell_size
-            + (self._rows - 1) * self._spacing
-        )
+        used_h = self._rows * self._cell_size + (self._rows - 1) * self._spacing
         return max(0, (self.height() - used_h) // 2)
 
     def _computeSnakePositions(self) -> List[QPoint]:
@@ -196,10 +187,7 @@ class _ThumbnailGridWidget(QWidget):
         for i in range(self._rows * self._cols):
             row = i // self._cols
             col_in_row = i % self._cols
-            if row % 2 == 0:
-                col = col_in_row
-            else:
-                col = self._cols - 1 - col_in_row
+            col = col_in_row if row % 2 == 0 else self._cols - 1 - col_in_row
             x = col * (self._cell_size + self._spacing)
             y = y_offset + row * (self._cell_size + self._spacing)
             positions.append(QPoint(x, y))
@@ -215,8 +203,7 @@ class _ThumbnailGridWidget(QWidget):
 
         Cells at indices below ``shift_count`` stay in place (they
         will be repainted by ``populate`` after the animation).
-        All other cells slide ``shift_count`` positions toward
-        index 0.
+        All other cells slide ``shift_count`` positions toward index 0.
         """
         if index < shift_count:
             return positions[index]
