+++ src/le_beta_vis/frontend/livemode/widgets/_ScaleGradientWidget.py
