+++ src/le_beta_vis/frontend/livemode/widgets/__init__.py
@@ -1,0 +1,7 @@
+from .DetectedClusterCollectionWidget import DetectedClusterCollectionWidget
+from .FeaturedClusterWidget import FeaturedClusterWidget
+
+__all__ = [
+    "DetectedClusterCollectionWidget",
+    "FeaturedClusterWidget",
+]
